Break into multiple files.
- Condition
- Predicate
- ServerCode
