+++ ThingIFSDK/ThingIFSDK/Trigger.swift
@@ -174,158 +174,6 @@ public class Trigger: NSObject, NSCoding {
 
 }
 
-/** Class represents Predicate */
-public class Predicate : NSObject, NSCoding {
-
-    public override init() {
-        super.init();
-    }
-
-    public required init(coder aDecoder: NSCoder) {
-        super.init();
-    }
-
-    public func encodeWithCoder(aCoder: NSCoder) {
-    }
-
-    /** Get Predicate as NSDictionary instance
-
-    - Returns: a NSDictionary instance
-    */
-    public func toNSDictionary() -> NSDictionary{
-        return NSDictionary()
-    }
-}
-
-/** Class represents Condition */
-public class Condition : NSObject, NSCoding {
-    public let clause: Clause!
-
-    /** Init Condition with Clause
-
-    - Parameter clause: Clause instance
-    */
-    public init(clause:Clause) {
-        self.clause = clause
-    }
-
-    public required init(coder aDecoder: NSCoder) {
-        self.clause = aDecoder.decodeObjectForKey("clause") as! Clause
-        super.init();
-    }
-
-    public func encodeWithCoder(aCoder: NSCoder) {
-        aCoder.encodeObject(self.clause, forKey: "clause")
-    }
-
-    /** Get Condition as NSDictionary instance
-
-    - Returns: a NSDictionary instance
-    */
-    public func toNSDictionary() -> NSDictionary {
-        return self.clause.toNSDictionary()
-    }
-
-    class func conditionWithNSDict(conditionDict: NSDictionary) -> Condition?{
-        if let clause = Condition.clauseWithNSDict(conditionDict) {
-            return Condition(clause: clause)
-        }else {
-            return nil
-        }
-    }
-
-    class func clauseWithNSDict(clauseDict: NSDictionary) -> Clause?{
-        var clause: Clause?
-        if let type = clauseDict["type"] as? String {
-            switch type {
-            case "range":
-                if let upperLimitNumber = clauseDict["upperLimit"] as? NSNumber, lowerLimitNumber = clauseDict["lowerLimit"] as? NSNumber, field = clauseDict["field"] as? String {
-                    if let upperIncluded = clauseDict["upperIncluded"] as? Bool, lowerIncluded = clauseDict["lowerIncluded"] as? Bool {
-                        if upperLimitNumber.isInt(){
-                            clause = RangeClause(field: field, lowerLimitInt: lowerLimitNumber.integerValue, lowerIncluded: lowerIncluded, upperLimit: upperLimitNumber.integerValue, upperIncluded: upperIncluded)
-                        }else if upperLimitNumber.isDouble() {
-                            clause = RangeClause(field: field, lowerLimitDouble: lowerLimitNumber.doubleValue, lowerIncluded: lowerIncluded, upperLimit: upperLimitNumber.doubleValue, upperIncluded: upperIncluded)
-                        }
-                    }
-                    break
-                }
-
-                if let upperLimitNumber = clauseDict["upperLimit"] as? NSNumber,
-                    filed = clauseDict["field"] as? String {
-                    if let upperIncluded = clauseDict["upperIncluded"] as? Bool {
-                        if upperLimitNumber.isInt(){
-                            clause = RangeClause(field: filed, upperLimitInt: upperLimitNumber.integerValue, upperIncluded: upperIncluded)
-                        }else if upperLimitNumber.isDouble() {
-                            clause = RangeClause(field: filed, upperLimitDouble: upperLimitNumber.doubleValue, upperIncluded: upperIncluded)
-                        }
-                    }
-                    break
-                }
-
-                if let lowerLimitNumber = clauseDict["lowerLimit"] as? NSNumber,
-                    filed = clauseDict["field"] as? String {
-                    if let lowerIncluded = clauseDict["lowerIncluded"] as? Bool {
-                        if lowerLimitNumber.isInt() {
-                            clause = RangeClause(field: filed, lowerLimitInt: lowerLimitNumber.integerValue, lowerIncluded: lowerIncluded)
-                        }else if lowerLimitNumber.isDouble() {
-                            clause = RangeClause(field: filed, lowerLimitDouble: lowerLimitNumber.doubleValue, lowerIncluded: lowerIncluded)
-                        }
-                    }
-                    break
-                }
-                break
-
-            case "eq":
-                if let field = clauseDict["field"] as? String, value = clauseDict["value"] {
-                    if value is String {
-                        clause = EqualsClause(field: field, stringValue: value as! String)
-                    }else if value is NSNumber {
-                        let numberValue = value as! NSNumber
-                        if numberValue.isBool() {
-                            clause = EqualsClause(field: field, boolValue: numberValue.boolValue)
-                        }else {
-                            clause = EqualsClause(field: field, intValue: numberValue.integerValue)
-                        }
-                    }
-                }
-
-            case "not":
-                if let clauseDict = clauseDict["clause"] as? NSDictionary {
-                    if let equalClause = Condition.clauseWithNSDict(clauseDict) as? EqualsClause{
-                        clause = NotEqualsClause(equalStmt: equalClause)
-                    }
-                }
-
-            case "and":
-                let andClause = AndClause()
-                if let clauseDicts = clauseDict["clauses"] as? NSArray {
-                    for clauseDict in clauseDicts {
-                        if let subClause = Condition.clauseWithNSDict(clauseDict as! NSDictionary) {
-                            andClause.add(subClause)
-                        }
-                    }
-                }
-                clause = andClause
-
-            case "or":
-                let orClause = OrClause()
-                if let clauseDicts = clauseDict["clauses"] as? NSArray {
-                    for clauseDict in clauseDicts {
-                        if let subClause = Condition.clauseWithNSDict(clauseDict as! NSDictionary) {
-                            orClause.add(subClause)
-                        }
-                    }
-                }
-                clause = orClause
-
-            default:
-                break
-            }
-        }
-        return clause
-    }
-}
-
 /** Enum defines when the Trigger is fired based on StatePredicate */
 public enum TriggersWhen : String {
     /* NOTE: These string values must not be changed. These values are
@@ -355,175 +203,3 @@ enum EventSource: String {
     case Schedule = "SCHEDULE"
 
 }
-
-/** Class represents SchedulePredicate. It is not supported now.*/
-public class SchedulePredicate: Predicate {
-    /** Specified schedule. (cron tab format) */
-    public let schedule: String
-
-    /** Instantiate new SchedulePredicate.
-
-    -Parameter schedule: Specify schedule. (cron tab format)
-    */
-    public init(schedule: String) {
-        self.schedule = schedule
-        super.init()
-    }
-
-    public required init(coder aDecoder: NSCoder) {
-        self.schedule = aDecoder.decodeObjectForKey("schedule") as! String;
-        super.init(coder: aDecoder);
-    }
-
-    public override func encodeWithCoder(aCoder: NSCoder) {
-        aCoder.encodeObject(self.schedule, forKey: "schedule");
-    }
-
-    /** Get Json object of SchedulePredicate instance
-
-    - Returns: Json object as an instance of NSDictionary
-    */
-    public override func toNSDictionary() -> NSDictionary {
-        return NSDictionary(dictionary: ["eventSource": EventSource.Schedule.rawValue, "schedule":self.schedule])
-    }
-}
-
-/** Class represents StatePredicate */
-public class StatePredicate: Predicate {
-    public let triggersWhen: TriggersWhen!
-    public let condition: Condition!
-
-    /** Initialize StatePredicate with Condition and TriggersWhen
-
-    - Parameter condition: Condition of the Trigger.
-    - Parameter triggersWhen: Specify TriggersWhen.
-    */
-    public init(condition:Condition, triggersWhen:TriggersWhen) {
-        self.triggersWhen = triggersWhen
-        self.condition = condition
-        super.init();
-    }
-
-    public required init(coder aDecoder: NSCoder) {
-        self.triggersWhen = TriggersWhen(rawValue: aDecoder.decodeObjectForKey("triggersWhen") as! String);
-        self.condition = aDecoder.decodeObjectForKey("condition") as! Condition;
-        super.init(coder: aDecoder);
-    }
-
-    public override func encodeWithCoder(aCoder: NSCoder) {
-        aCoder.encodeObject(self.triggersWhen.rawValue, forKey: "triggersWhen");
-        aCoder.encodeObject(self.condition, forKey: "condition");
-    }
-
-    /** Get StatePredicate as NSDictionary instance
-
-    - Returns: a NSDictionary instance
-    */
-    public override func toNSDictionary() -> NSDictionary {
-        return NSDictionary(dictionary: ["eventSource": EventSource.States.rawValue, "triggersWhen": self.triggersWhen.rawValue, "condition": self.condition.toNSDictionary()])
-    }
-
-    class func statePredicateWithNSDict(predicateDict: NSDictionary) -> StatePredicate?{
-        var triggersWhen: TriggersWhen?
-        var condition: Condition?
-        if let triggersWhenString = predicateDict["triggersWhen"] as? String {
-            triggersWhen = TriggersWhen(rawValue: triggersWhenString)
-        }
-
-        if let conditionDict = predicateDict["condition"] as? NSDictionary {
-            condition = Condition.conditionWithNSDict(conditionDict)
-        }
-
-        if triggersWhen != nil && condition != nil {
-            return StatePredicate(condition: condition!, triggersWhen: triggersWhen!)
-        }else {
-            return nil
-        }
-    }
-}
-
-public class ServerCode : NSObject, NSCoding {
-    // MARK: - Implements NSCoding protocol
-    public func encodeWithCoder(aCoder: NSCoder) {
-        aCoder.encodeObject(self.endpoint, forKey: "endpoint")
-        aCoder.encodeObject(self.executorAccessToken, forKey: "executorAccessToken")
-        aCoder.encodeObject(self.targetAppID, forKey: "targetAppID")
-        aCoder.encodeObject(self.parameters, forKey: "parameters")
-    }
-    
-    // MARK: - Implements NSCoding protocol
-    public required init(coder aDecoder: NSCoder) {
-        self.endpoint = aDecoder.decodeObjectForKey("endpoint") as! String
-        self.executorAccessToken = aDecoder.decodeObjectForKey("executorAccessToken") as? String
-        self.targetAppID = aDecoder.decodeObjectForKey("targetAppID") as? String
-        self.parameters = aDecoder.decodeObjectForKey("parameters") as? Dictionary<String, AnyObject>
-    }
-    
-    /** Endpoint to call on servercode */
-    public var endpoint: String
-    /** This token will be used to call the external appID endpoint */
-    public var executorAccessToken: String?
-    /** If provided, servercode endpoint will be called for this appid. Otherwise same appID of trigger is used */
-    public var targetAppID: String?
-    /** Parameters to pass to the servercode function */
-    public var parameters: Dictionary<String, AnyObject>?
-
-    /** Init TriggeredServerCodeResult with necessary attributes
-     
-     - Parameter endpoint: Endpoint to call on servercode
-     - Parameter executorAccessToken: This token will be used to call the external appID endpoint
-     - Parameter targetAppID: If provided, servercode endpoint will be called for this appid. Otherwise same appID of trigger is used
-     - Parameter parameters: Parameters to pass to the servercode function
-     */
-    public init(endpoint: String, executorAccessToken: String?, targetAppID: String?, parameters: Dictionary<String, AnyObject>?) {
-        self.endpoint = endpoint
-        self.executorAccessToken = executorAccessToken
-        self.targetAppID = targetAppID
-        self.parameters = parameters
-    }
-
-    func toNSDictionary() -> NSDictionary {
-        let dict = NSMutableDictionary(dictionary: ["endpoint": self.endpoint])
-        if self.executorAccessToken != nil {
-            dict["executorAccessToken"] = self.executorAccessToken
-        }
-        if self.targetAppID != nil {
-            dict["targetAppID"] = self.targetAppID
-        }
-        if self.parameters != nil {
-            dict["parameters"] = self.parameters
-        }
-        return dict
-    }
-
-    public override func isEqual(object: AnyObject?) -> Bool {
-        guard let aServerCode = object as? ServerCode else{
-            return false
-        }
-        if self.parameters == nil || aServerCode.parameters == nil {
-            if self.parameters == nil && aServerCode.parameters == nil {
-                return true
-            }
-            return false
-        }
-        return self.endpoint == aServerCode.endpoint &&
-            self.executorAccessToken == aServerCode.executorAccessToken &&
-            self.targetAppID == aServerCode.targetAppID &&
-            NSDictionary(dictionary: self.parameters!).isEqualToDictionary(aServerCode.parameters!)
-        
-    }
-    
-    class func serverCodeWithNSDictionary(nsDict: NSDictionary!) -> ServerCode?{
-        let endpoint = nsDict["endpoint"] as? String
-        let executorAccessToken = nsDict["executorAccessToken"] as? String
-        let targetAppID = nsDict["targetAppID"] as? String
-        let parameters = nsDict["parameters"] as? Dictionary<String, AnyObject>
-        var serverCode: ServerCode?
-        if (endpoint != nil) {
-            serverCode = ServerCode(endpoint:endpoint!, executorAccessToken:executorAccessToken, targetAppID:targetAppID, parameters:parameters)
-        }
-        return serverCode
-    }
-
-}
-
